Refactor validation to utility module and enhance tests
Move validation logic to retry.py for better cohesion. Enforce
validation in both trigger constructors within databricks.py.
Add direct trigger regression tests in test_databricks.py and
update sensor test setup to maintain deferrable branch coverage.
