+++ providers/databricks/src/airflow/providers/databricks/triggers/databricks.py
@@ -23,6 +23,7 @@ from typing import Any
 
 from airflow.providers.databricks.hooks.databricks import DatabricksHook
 from airflow.providers.databricks.utils.databricks import extract_failed_task_errors_async
+from airflow.providers.databricks.utils.retry import validate_deferrable_databricks_retry_args
 from airflow.triggers.base import BaseTrigger, TriggerEvent
 
 
@@ -55,6 +56,7 @@ class DatabricksExecutionTrigger(BaseTrigger):
         caller: str = "DatabricksExecutionTrigger",
     ) -> None:
         super().__init__()
+        validate_deferrable_databricks_retry_args(retry_args, owner=self.__class__.__name__)
         self.run_id = run_id
         self.databricks_conn_id = databricks_conn_id
         self.polling_period_seconds = polling_period_seconds
@@ -151,6 +153,7 @@ class DatabricksSQLStatementExecutionTrigger(BaseTrigger):
         caller: str = "DatabricksSQLStatementExecutionTrigger",
     ) -> None:
         super().__init__()
+        validate_deferrable_databricks_retry_args(retry_args, owner=self.__class__.__name__)
         self.statement_id = statement_id
         self.databricks_conn_id = databricks_conn_id
         self.end_time = end_time
