+++ providers/databricks/tests/unit/databricks/triggers/test_databricks.py
@@ -21,6 +21,7 @@ import time
 from unittest import mock
 
 import pytest
+from tenacity import wait_incrementing
 
 from airflow.models import Connection
 from airflow.providers.databricks.hooks.databricks import RunState, SQLStatementState
@@ -166,6 +167,16 @@ class TestDatabricksExecutionTrigger:
         _, kwargs = trigger.serialize()
         restored = DatabricksExecutionTrigger(**kwargs)
         assert restored.caller == CALLER
+
+    def test_init_rejects_non_serializable_retry_args(self):
+        with pytest.raises(
+            ValueError, match="does not support non-serializable databricks_retry_args when deferrable=True"
+        ):
+            DatabricksExecutionTrigger(
+                run_id=RUN_ID,
+                databricks_conn_id=DEFAULT_CONN_ID,
+                retry_args={"wait": wait_incrementing(start=1, increment=1, max=3)},
+            )
 
     @pytest.mark.asyncio
     @mock.patch("airflow.providers.databricks.hooks.databricks.DatabricksHook.a_get_run_output")
@@ -325,6 +336,17 @@ class TestDatabricksSQLStatementExecutionTrigger:
         _, kwargs = trigger.serialize()
         restored = DatabricksSQLStatementExecutionTrigger(**kwargs)
         assert restored.caller == CALLER
+
+    def test_init_rejects_non_serializable_retry_args(self):
+        with pytest.raises(
+            ValueError, match="does not support non-serializable databricks_retry_args when deferrable=True"
+        ):
+            DatabricksSQLStatementExecutionTrigger(
+                statement_id=STATEMENT_ID,
+                databricks_conn_id=DEFAULT_CONN_ID,
+                end_time=self.end_time,
+                retry_args={"wait": wait_incrementing(start=1, increment=1, max=3)},
+            )
 
     @pytest.mark.asyncio
     @mock.patch("airflow.providers.databricks.hooks.databricks.DatabricksHook.a_get_sql_statement_state")
